requirements.txt
aiogram==2.14.2
aiohttp==3.7.4.post0
async-timeout==3.0.1
google-api-core==1.31.0
google-api-python-client==2.14.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
oauthlib==3.1.1
pip==21.1.3
requests==2.25.1
requests-oauthlib==1.3.0
TikTokApi==3.9.9.1
urllib3==1.26.5
websockets==9.1
yarl==1.6.3
youtube-dl==2021.6.6
